// lib/api/types.ts
export interface AdminProduct {
  id: number
  slug: string
  name: string
  category: string
  section: string
  price: number
  originalPrice: number | null
  stock: number
  tag: string
  badge: string
  active: boolean
  featured: boolean
  externalProductType: string
  externalAccountType: string
  supplier?: string
  bulkmailProductId?: number | null
}